+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for NeoFS Node
 ### Fixed
 
 - Losing request context in eACL response checks (#1595)
+- Do not require `lifetime` flag in `session create` CLI command (#1655)
 
 ### Removed